Switch to Fedora for nettest and use packaged netperf
netperf currently fails to build with the latest Alpine images, even
with _GNU_SOURCE fixes.

To allow building the nettest image again, this switches to Fedora,
where netperf is available as a package (so we don't need to deal with
building it). The image size is more than doubled, from 24.2MiB with
Alpine to 59.8MiB with Fedora.

This ends up with nmap ncat rather than busybox ncat, so the nettest
scripts are adjusted to suit, and a busybox nc emulation script is
included for other users of nc.

Signed-off-by: Stephen Kitt <skitt@redhat.com>